jd_service/requirements.txt
fastapi
uvicorn
aiofiles
sqlalchemy
psycopg2-binary
pydantic
kafka-python
python-multipart
pdfminer.six
python-docx
